Fix to #10284 - EF Core does not cast from Nullable to underlying type
Problem was that we were too aggressive in applying null compensation during navigation rewrite. This could lead to type change of an argument to a function - if the argument was a non-nullable property accessed via optional navigation). This in turn causes issues during compilation, because new method arguments must match method signature.

Fix is to unwrap null compensation for arguments of the function.